+++ GDX.Tests/EditMode/ReflectionTests.cs
@@ -13,11 +13,14 @@ namespace GDX
     /// </summary>
     public class ReflectionTests
     {
+        const string k_ProjectBrowserFullName = "UnityEditor.ProjectBrowser";
+        const string k_BadTypeFullName = "UnityEditor.ProjectBrowserBad";
+
         [Test]
         [Category(Core.TestCategory)]
         public void GetDefault_Nullable_ReturnsNull()
         {
-            Assert.IsTrue(typeof(EditorWindow).GetDefault() == null, "A null value was expected.");
+            Assert.IsTrue(typeof(EditorWindow).GetDefault() == null);
         }
 
         [Test]
@@ -26,30 +29,30 @@ namespace GDX
         {
             object defaultBoxedValue = typeof(int).GetDefault();
             int defaultValue = (int)defaultBoxedValue;
-            Assert.IsTrue(defaultValue == 0,  "A non-null value was expected.");
+            Assert.IsTrue(defaultValue == 0);
         }
 
         [Test]
         [Category(Core.TestCategory)]
         public void GetType_QualifiedType_ReturnsType()
         {
-            System.Type projectBrowser = Reflection.GetType("UnityEditor.ProjectBrowser");
-            Assert.IsTrue(projectBrowser != null, "Expected to be able to find UnityEditor.ProjectBrowser type.");
+            System.Type projectBrowser = Reflection.GetType(k_ProjectBrowserFullName);
+            Assert.IsTrue(projectBrowser != null);
         }
 
         [Test]
         [Category(Core.TestCategory)]
         public void GetType_BadType_ReturnsNull()
         {
-            System.Type projectBrowser = Reflection.GetType("UnityEditor.ProjectBrowserBad");
-            Assert.IsTrue(projectBrowser == null, "Expected null type for bad type.");
+            System.Type projectBrowser = Reflection.GetType(k_BadTypeFullName);
+            Assert.IsTrue(projectBrowser == null);
         }
 
         [Test]
         [Category(Core.TestCategory)]
         public void InvokeStaticMethod_ProjectBrowser_ReturnsListOfBrowsers()
         {
-            object response = Reflection.InvokeStaticMethod("UnityEditor.ProjectBrowser", "GetAllProjectBrowsers");
+            object response = Reflection.InvokeStaticMethod(k_ProjectBrowserFullName, "GetAllProjectBrowsers");
             Assert.IsTrue(response != null);
         }
 
@@ -58,7 +61,7 @@ namespace GDX
         [Category(Core.TestCategory)]
         public void InvokeStaticMethod_BadMethod_ReturnsNull()
         {
-            object response = Reflection.InvokeStaticMethod("UnityEditor.ProjectBrowser", "GetAllProjectBrowsersBad");
+            object response = Reflection.InvokeStaticMethod(k_ProjectBrowserFullName, "GetAllProjectBrowsersBad");
             Assert.IsTrue(response == null);
         }
 
@@ -66,7 +69,7 @@ namespace GDX
         [Category(Core.TestCategory)]
         public void InvokeStaticMethod_BadType_ReturnsNull()
         {
-            object response = Reflection.InvokeStaticMethod("UnityEditor.ProjectBrowserBad", "GetAllProjectBrowsersBad");
+            object response = Reflection.InvokeStaticMethod(k_BadTypeFullName, "GetAllProjectBrowsersBad");
             Assert.IsTrue(response == null);
         }
 
@@ -76,11 +79,11 @@ namespace GDX
         [Category(Core.TestCategory)]
         public void SetFieldOrPropertyValue_ProjectBrowser_SetIsLocked()
         {
-            System.Type projectBrowser = Reflection.GetType("UnityEditor.ProjectBrowser");
-            Assert.IsTrue(projectBrowser != null, "Expected to be able to find UnityEditor.ProjectBrowser type.");
+            System.Type projectBrowser = Reflection.GetType(k_ProjectBrowserFullName);
+            Assert.IsTrue(projectBrowser != null);
 
             EditorWindow projectBrowserWindow = EditorWindow.GetWindow(projectBrowser);
-            Assert.IsTrue(projectBrowserWindow != null, "Expected reference to ProjectBrowser.");
+            Assert.IsTrue(projectBrowserWindow != null);
 
             Assert.IsTrue(Reflection.TryGetFieldOrPropertyValue(projectBrowserWindow, "isLocked", out object previous));
 
@@ -98,11 +101,11 @@ namespace GDX
         [Category(Core.TestCategory)]
         public void SetFieldOrPropertyValue_ProjectBrowser_SetSearchFieldText()
         {
-            System.Type projectBrowser = Reflection.GetType("UnityEditor.ProjectBrowser");
-            Assert.IsTrue(projectBrowser != null, "Expected to be able to find UnityEditor.ProjectBrowser type.");
+            System.Type projectBrowser = Reflection.GetType(k_ProjectBrowserFullName);
+            Assert.IsTrue(projectBrowser != null);
 
             EditorWindow projectBrowserWindow = EditorWindow.GetWindow(projectBrowser);
-            Assert.IsTrue(projectBrowserWindow != null, "Expected reference to ProjectBrowser.");
+            Assert.IsTrue(projectBrowserWindow != null);
 
             Assert.IsTrue(Reflection.SetFieldOrPropertyValue(
                 projectBrowserWindow, "m_SearchFieldText", "test"));
@@ -119,34 +122,32 @@ namespace GDX
         [Category(Core.TestCategory)]
         public void SetFieldOrPropertyValue_NoType_ReturnsFalse()
         {
-            System.Type projectBrowser = Reflection.GetType("UnityEditor.ProjectBrowser");
-            Assert.IsTrue(projectBrowser != null, "Expected to be able to find UnityEditor.ProjectBrowser type.");
+            System.Type projectBrowser = Reflection.GetType(k_ProjectBrowserFullName);
+            Assert.IsTrue(projectBrowser != null);
 
             EditorWindow projectBrowserWindow = EditorWindow.GetWindow(projectBrowser);
-            Assert.IsTrue(projectBrowserWindow != null, "Expected reference to ProjectBrowser.");
+            Assert.IsTrue(projectBrowserWindow != null);
 
             Assert.IsFalse(Reflection.SetFieldOrPropertyValue(projectBrowserWindow,
-                "m_SearchFieldTextBad", "test"),
-                "Expected default value for non existing type.");
+                "m_SearchFieldTextBad", "test"));
         }
 
         [Test]
         [Category(Core.TestCategory)]
         public void SetFieldOrPropertyValue_NoTarget_ReturnsFalse()
         {
-            Assert.IsFalse(Reflection.SetFieldOrPropertyValue(null, "m_SearchFieldTextBad", "test"),
-                "Expected false value when no target.");
+            Assert.IsFalse(Reflection.SetFieldOrPropertyValue(null, "m_SearchFieldTextBad", "test"));
         }
 
         [Test]
         [Category(Core.TestCategory)]
         public void SetFieldValue_ProjectBrowser_SetSearchFieldText()
         {
-            System.Type projectBrowser = Reflection.GetType("UnityEditor.ProjectBrowser");
-            Assert.IsTrue(projectBrowser != null, "Expected to be able to find UnityEditor.ProjectBrowser type.");
+            System.Type projectBrowser = Reflection.GetType(k_ProjectBrowserFullName);
+            Assert.IsTrue(projectBrowser != null);
 
             EditorWindow projectBrowserWindow = EditorWindow.GetWindow(projectBrowser);
-            Assert.IsTrue(projectBrowserWindow != null, "Expected reference to ProjectBrowser.");
+            Assert.IsTrue(projectBrowserWindow != null);
 
             Assert.IsTrue(Reflection.SetFieldValue(
                 projectBrowserWindow, projectBrowser, "m_SearchFieldText", "test"));
@@ -163,51 +164,47 @@ namespace GDX
         [Category(Core.TestCategory)]
         public void SetFieldValue_NoType_ReturnsFalse()
         {
-            System.Type projectBrowser = Reflection.GetType("UnityEditor.ProjectBrowser");
-            Assert.IsTrue(projectBrowser != null, "Expected to be able to find UnityEditor.ProjectBrowser type.");
+            System.Type projectBrowser = Reflection.GetType(k_ProjectBrowserFullName);
+            Assert.IsTrue(projectBrowser != null);
 
             EditorWindow projectBrowserWindow = EditorWindow.GetWindow(projectBrowser);
-            Assert.IsTrue(projectBrowserWindow != null, "Expected reference to ProjectBrowser.");
+            Assert.IsTrue(projectBrowserWindow != null);
 
             Assert.IsFalse(Reflection.SetFieldValue(projectBrowserWindow,
-                    null, "m_SearchFieldTextBad", "test"),
-                "Expected false for non existing type.");
+                    null, "m_SearchFieldTextBad", "test"));
         }
 
         [Test]
         [Category(Core.TestCategory)]
         public void SetFieldValue_BadField_ReturnsFalse()
         {
-            System.Type projectBrowser = Reflection.GetType("UnityEditor.ProjectBrowser");
-            Assert.IsTrue(projectBrowser != null, "Expected to be able to find UnityEditor.ProjectBrowser type.");
+            System.Type projectBrowser = Reflection.GetType(k_ProjectBrowserFullName);
+            Assert.IsTrue(projectBrowser != null);
 
             EditorWindow projectBrowserWindow = EditorWindow.GetWindow(projectBrowser);
-            Assert.IsTrue(projectBrowserWindow != null, "Expected reference to ProjectBrowser.");
+            Assert.IsTrue(projectBrowserWindow != null);
 
             Assert.IsFalse(Reflection.SetFieldValue(projectBrowserWindow,
-                    projectBrowser, "m_SearchFieldTextBad", "test"),
-                "Expected false for non existing field.");
+                    projectBrowser, "m_SearchFieldTextBad", "test"));
         }
 
         [Test]
         [Category(Core.TestCategory)]
         public void SetPropertyValue_QualifiedProperty_IsSet()
         {
-            System.Type projectBrowser = Reflection.GetType("UnityEditor.ProjectBrowser");
-            Assert.IsTrue(projectBrowser != null, "Expected to be able to find UnityEditor.ProjectBrowser type.");
+            System.Type projectBrowser = Reflection.GetType(k_ProjectBrowserFullName);
+            Assert.IsTrue(projectBrowser != null);
             EditorWindow projectBrowserWindow = EditorWindow.GetWindow(projectBrowser);
-            Assert.IsTrue(projectBrowserWindow != null, "Expected reference to ProjectBrowser.");
+            Assert.IsTrue(projectBrowserWindow != null);
 
             Assert.IsTrue(
-                Reflection.TryGetPropertyValue(projectBrowserWindow, projectBrowser, "isLocked", out bool previousValue),
-                "Expected to be able to retrieve the value.");
+                Reflection.TryGetPropertyValue(projectBrowserWindow, projectBrowser, "isLocked", out bool previousValue));
 
             Assert.IsTrue(
-                Reflection.SetPropertyValue(projectBrowserWindow, projectBrowser, "isLocked", !previousValue),
-                "Expected to be able to set the value.");
+                Reflection.SetPropertyValue(projectBrowserWindow, projectBrowser, "isLocked", !previousValue));
 
             Reflection.TryGetPropertyValue(projectBrowserWindow, projectBrowser, "isLocked", out bool newValue);
-            Assert.IsTrue(!previousValue == newValue, "Expected set value to match retried value.");
+            Assert.IsTrue(!previousValue == newValue);
 
             Reflection.SetPropertyValue(projectBrowserWindow, projectBrowser, "isLocked", previousValue);
         }
@@ -216,28 +213,26 @@ namespace GDX
         [Category(Core.TestCategory)]
         public void SetPropertyValue_NoType_IsFalse()
         {
-            System.Type projectBrowser = Reflection.GetType("UnityEditor.ProjectBrowser");
-            Assert.IsTrue(projectBrowser != null, "Expected to be able to find UnityEditor.ProjectBrowser type.");
+            System.Type projectBrowser = Reflection.GetType(k_ProjectBrowserFullName);
+            Assert.IsTrue(projectBrowser != null);
             EditorWindow projectBrowserWindow = EditorWindow.GetWindow(projectBrowser);
-            Assert.IsTrue(projectBrowserWindow != null, "Expected reference to ProjectBrowser.");
+            Assert.IsTrue(projectBrowserWindow != null);
 
             Assert.IsFalse(
-                Reflection.SetPropertyValue(projectBrowserWindow, null, "isLocked", false),
-                "Expected to not be able to set the value.");
+                Reflection.SetPropertyValue(projectBrowserWindow, null, "isLocked", false));
         }
 
         [Test]
         [Category(Core.TestCategory)]
         public void SetPropertyValue_InvalidName_IsFalse()
         {
-            System.Type projectBrowser = Reflection.GetType("UnityEditor.ProjectBrowser");
-            Assert.IsTrue(projectBrowser != null, "Expected to be able to find UnityEditor.ProjectBrowser type.");
+            System.Type projectBrowser = Reflection.GetType(k_ProjectBrowserFullName);
+            Assert.IsTrue(projectBrowser != null);
             EditorWindow projectBrowserWindow = EditorWindow.GetWindow(projectBrowser);
-            Assert.IsTrue(projectBrowserWindow != null, "Expected reference to ProjectBrowser.");
+            Assert.IsTrue(projectBrowserWindow != null);
 
             Assert.IsFalse(
-                Reflection.SetPropertyValue(projectBrowserWindow, projectBrowser, "isLockedBad", false),
-                "Expected to not be able to set the value.");
+                Reflection.SetPropertyValue(projectBrowserWindow, projectBrowser, "isLockedBad", false));
         }
 
         [Test]
@@ -245,156 +240,153 @@ namespace GDX
         public void TryGetFieldOrPropertyValue_NoTarget_ReturnsDefault()
         {
             Assert.IsFalse(Reflection.TryGetFieldOrPropertyValue(
-                null, "UnityEditor.ProjectBrowser", out object missedObject));
-            Assert.IsNull(missedObject, "Expected null returned object with no target.");
+                null, k_ProjectBrowserFullName, out object missedObject));
+            Assert.IsNull(missedObject);
         }
 
         [Test]
         [Category(Core.TestCategory)]
         public void TryGetFieldOrPropertyValue_NoField_ReturnsDefault()
         {
-            System.Type projectBrowser = Reflection.GetType("UnityEditor.ProjectBrowser");
-            Assert.IsTrue(projectBrowser != null, "Expected to be able to find UnityEditor.ProjectBrowser type.");
+            System.Type projectBrowser = Reflection.GetType(k_ProjectBrowserFullName);
+            Assert.IsTrue(projectBrowser != null);
 
             EditorWindow projectBrowserWindow = EditorWindow.GetWindow(projectBrowser);
-            Assert.IsTrue(projectBrowserWindow != null, "Expected reference to ProjectBrowser");
+            Assert.IsTrue(projectBrowserWindow != null);
 
             Assert.IsFalse(Reflection.TryGetFieldOrPropertyValue(
-                projectBrowserWindow, "UnityEditor.ProjectBrowserBad", out object missedObject));
-            Assert.IsNull(missedObject, "Expected null returned object with a bad field or property.");
+                projectBrowserWindow, k_BadTypeFullName, out object missedObject));
+            Assert.IsNull(missedObject);
         }
 
         [Test]
         [Category(Core.TestCategory)]
         public void TryGetFieldOrPropertyValue_BadField_ReturnsStartGridSize()
         {
-            System.Type projectBrowser = Reflection.GetType("UnityEditor.ProjectBrowser");
-            Assert.IsTrue(projectBrowser != null, "Expected to be able to find UnityEditor.ProjectBrowser type.");
+            System.Type projectBrowser = Reflection.GetType(k_ProjectBrowserFullName);
+            Assert.IsTrue(projectBrowser != null);
 
             EditorWindow projectBrowserWindow = EditorWindow.GetWindow(projectBrowser);
-            Assert.IsTrue(projectBrowserWindow != null, "Expected reference to ProjectBrowser");
+            Assert.IsTrue(projectBrowserWindow != null);
 
             Assert.IsTrue(Reflection.TryGetFieldOrPropertyValue(
                 projectBrowserWindow, "m_StartGridSize", out object gridSize));
 
-            Assert.IsTrue(gridSize != null, "Expected non-null value for m_StartGridSize.");
+            Assert.IsTrue(gridSize != null);
         }
 
         [Test]
         [Category(Core.TestCategory)]
         public void TryGetFieldOrPropertyValue_ProjectBrowserField_ReturnsStartGridSize()
         {
-            System.Type projectBrowser = Reflection.GetType("UnityEditor.ProjectBrowser");
-            Assert.IsTrue(projectBrowser != null, "Expected to be able to find UnityEditor.ProjectBrowser type.");
+            System.Type projectBrowser = Reflection.GetType(k_ProjectBrowserFullName);
+            Assert.IsTrue(projectBrowser != null);
 
             EditorWindow projectBrowserWindow = EditorWindow.GetWindow(projectBrowser);
-            Assert.IsTrue(projectBrowserWindow != null, "Expected reference to ProjectBrowser");
+            Assert.IsTrue(projectBrowserWindow != null);
 
             Assert.IsTrue(Reflection.TryGetFieldOrPropertyValue(
                 projectBrowserWindow, "m_StartGridSize", out object gridSize));
 
-            Assert.IsTrue(gridSize != null, "Expected non-null value for m_StartGridSize.");
+            Assert.IsTrue(gridSize != null);
         }
 
         [Test]
         [Category(Core.TestCategory)]
         public void TryGetFieldOrPropertyValue_ProjectBrowserProperty_ReturnsIsLocked()
         {
-            System.Type projectBrowser = Reflection.GetType("UnityEditor.ProjectBrowser");
-            Assert.IsTrue(projectBrowser != null, "Expected to be able to find UnityEditor.ProjectBrowser type.");
+            System.Type projectBrowser = Reflection.GetType(k_ProjectBrowserFullName);
+            Assert.IsTrue(projectBrowser != null);
 
             EditorWindow projectBrowserWindow = EditorWindow.GetWindow(projectBrowser);
-            Assert.IsTrue(projectBrowserWindow != null, "Expected reference to ProjectBrowser.");
+            Assert.IsTrue(projectBrowserWindow != null);
 
             Assert.IsTrue(Reflection.TryGetFieldOrPropertyValue(projectBrowserWindow, "isLocked",
                 out object toggle));
 
-            Assert.IsTrue(toggle != null, "Expected non-null value for isLocked.");
+            Assert.IsTrue(toggle != null);
         }
 
         [Test]
         [Category(Core.TestCategory)]
         public void TryGetFieldValue_ProjectBrowser_ReturnsTrueGetFieldText()
         {
-            System.Type projectBrowser = Reflection.GetType("UnityEditor.ProjectBrowser");
-            Assert.IsTrue(projectBrowser != null, "Expected to be able to find UnityEditor.ProjectBrowser type.");
+            System.Type projectBrowser = Reflection.GetType(k_ProjectBrowserFullName);
+            Assert.IsTrue(projectBrowser != null);
 
             EditorWindow projectBrowserWindow = EditorWindow.GetWindow(projectBrowser);
-            Assert.IsTrue(projectBrowserWindow != null, "Expected reference to ProjectBrowser.");
+            Assert.IsTrue(projectBrowserWindow != null);
 
             Assert.IsTrue(Reflection.TryGetFieldValue(
                 projectBrowserWindow, projectBrowser, "m_SearchFieldText", out string query));
 
-            Assert.IsTrue(query != null, "Expected non-null value for m_SearchFieldText.");
+            Assert.IsTrue(query != null);
         }
 
         [Test]
         [Category(Core.TestCategory)]
         public void TryGetFieldValue_NoType_ReturnsDefault()
         {
-            System.Type projectBrowser = Reflection.GetType("UnityEditor.ProjectBrowser");
-            Assert.IsTrue(projectBrowser != null, "Expected to be able to find UnityEditor.ProjectBrowser type.");
+            System.Type projectBrowser = Reflection.GetType(k_ProjectBrowserFullName);
+            Assert.IsTrue(projectBrowser != null);
 
             EditorWindow projectBrowserWindow = EditorWindow.GetWindow(projectBrowser);
-            Assert.IsTrue(projectBrowserWindow != null, "Expected reference to ProjectBrowser.");
+            Assert.IsTrue(projectBrowserWindow != null);
 
             Assert.IsFalse(Reflection.TryGetFieldValue(
                 projectBrowserWindow, null, "m_SearchFieldTextBad", out string returnValue));
-            Assert.IsNull(returnValue, "Expected return value to be null.");
+            Assert.IsNull(returnValue);
         }
 
         [Test]
         [Category(Core.TestCategory)]
         public void TryGetFieldValue_BadField_ReturnsDefault()
         {
-            System.Type projectBrowser = Reflection.GetType("UnityEditor.ProjectBrowser");
-            Assert.IsTrue(projectBrowser != null, "Expected to be able to find UnityEditor.ProjectBrowser type.");
+            System.Type projectBrowser = Reflection.GetType(k_ProjectBrowserFullName);
+            Assert.IsTrue(projectBrowser != null);
 
             EditorWindow projectBrowserWindow = EditorWindow.GetWindow(projectBrowser);
-            Assert.IsTrue(projectBrowserWindow != null, "Expected reference to ProjectBrowser.");
+            Assert.IsTrue(projectBrowserWindow != null);
 
             Assert.IsFalse(Reflection.TryGetFieldValue(
                 projectBrowserWindow, projectBrowser, "m_SearchFieldTextBad", out string returnValue));
-            Assert.IsNull(returnValue, "Expected return value to be null.");
+            Assert.IsNull(returnValue);
         }
 
         [Test]
         [Category(Core.TestCategory)]
         public void TryGetPropertyValue_QualifiedProperty_IsTrue()
         {
-            System.Type projectBrowser = Reflection.GetType("UnityEditor.ProjectBrowser");
-            Assert.IsTrue(projectBrowser != null, "Expected to be able to find UnityEditor.ProjectBrowser type.");
+            System.Type projectBrowser = Reflection.GetType(k_ProjectBrowserFullName);
+            Assert.IsTrue(projectBrowser != null);
             EditorWindow projectBrowserWindow = EditorWindow.GetWindow(projectBrowser);
-            Assert.IsTrue(projectBrowserWindow != null, "Expected reference to ProjectBrowser.");
+            Assert.IsTrue(projectBrowserWindow != null);
 
-            Assert.IsTrue(Reflection.TryGetPropertyValue(projectBrowserWindow, projectBrowser, "isLocked", out bool _),
-                "Expected true response from method.");
+            Assert.IsTrue(Reflection.TryGetPropertyValue(projectBrowserWindow, projectBrowser, "isLocked", out bool _));
         }
 
         [Test]
         [Category(Core.TestCategory)]
         public void TryGetPropertyValue_NoType_IsFalse()
         {
-            System.Type projectBrowser = Reflection.GetType("UnityEditor.ProjectBrowser");
-            Assert.IsTrue(projectBrowser != null, "Expected to be able to find UnityEditor.ProjectBrowser type.");
+            System.Type projectBrowser = Reflection.GetType(k_ProjectBrowserFullName);
+            Assert.IsTrue(projectBrowser != null);
             EditorWindow projectBrowserWindow = EditorWindow.GetWindow(projectBrowser);
-            Assert.IsTrue(projectBrowserWindow != null, "Expected reference to ProjectBrowser.");
+            Assert.IsTrue(projectBrowserWindow != null);
 
-            Assert.IsFalse(Reflection.TryGetPropertyValue(projectBrowserWindow, null, "isLocked", out bool _),
-                "Expected false response from method.");
+            Assert.IsFalse(Reflection.TryGetPropertyValue(projectBrowserWindow, null, "isLocked", out bool _));
         }
 
         [Test]
         [Category(Core.TestCategory)]
         public void TryGetPropertyValue_InvalidName_IsFalse()
         {
-            System.Type projectBrowser = Reflection.GetType("UnityEditor.ProjectBrowser");
-            Assert.IsTrue(projectBrowser != null, "Expected to be able to find UnityEditor.ProjectBrowser type.");
+            System.Type projectBrowser = Reflection.GetType(k_ProjectBrowserFullName);
+            Assert.IsTrue(projectBrowser != null);
             EditorWindow projectBrowserWindow = EditorWindow.GetWindow(projectBrowser);
-            Assert.IsTrue(projectBrowserWindow != null, "Expected reference to ProjectBrowser.");
+            Assert.IsTrue(projectBrowserWindow != null);
 
-            Assert.IsFalse(Reflection.TryGetPropertyValue(projectBrowserWindow, projectBrowser, "isLockedBad", out bool _),
-                "Expected false response from method.");
+            Assert.IsFalse(Reflection.TryGetPropertyValue(projectBrowserWindow, projectBrowser, "isLockedBad", out bool _));
         }
     }
 }